+++ input.txt
@@ -5,3 +5,45 @@
 22.84,50.00
 0.01,1000.00
 0.03,0.05
+2.12,3.00
+1.97,2.00
+3.33,5.00
+12.47,15.00
+22.84,50.00
+0.01,1000.00
+0.03,0.05
+2.12,3.00
+1.97,2.00
+3.33,5.00
+12.47,15.00
+22.84,50.00
+0.01,1000.00
+0.03,0.05
+2.12,3.00
+1.97,2.00
+3.33,5.00
+12.47,15.00
+22.84,50.00
+0.01,1000.00
+0.03,0.05
+2.12,3.00
+1.97,2.00
+3.33,5.00
+12.47,15.00
+22.84,50.00
+0.01,1000.00
+0.03,0.05
+2.12,3.00
+1.97,2.00
+3.33,5.00
+12.47,15.00
+22.84,50.00
+0.01,1000.00
+0.03,0.05
+2.12,3.00
+1.97,2.00
+3.33,5.00
+12.47,15.00
+22.84,50.00
+0.01,1000.00
+0.03,0.05
